requirements.txt
torch==2.2.2
torchvision==0.17.2
pillow~=10.4.0
pycocotools~=2.0.8
jsonargparse==4.33.2